backend/lyalya/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14.0,<3.15.0
psycopg2-binary==2.9.9
model-bakery==1.20.4
django-cors-headers>=4.3.1,<4.4.0